# src/docs/swagger.yaml
basePath: /
definitions:
...
  carwise.BrandCreateRequest:
    description: Request body for creating a new brand
    properties:
      image_id:
        example: img_123456
        type: string
      name:
        example: BMW
        type: string
    required:
    - name
    type: object
  carwise.BrandUpdateRequest:
    description: Request body for updating an existing brand
    properties:
      image_id:
        example: img_123456
        type: string
      name:
        example: BMW
        type: string
    required:
    - name
    type: object
  carwise.BrandWithDetails:
    properties:
      id:
        type: string
      image_id:
        type: string
      name:
        type: string
      series:
        items:
          $ref: '#/definitions/carwise.SeriesDetail'
        type: array
    type: object
...
  carwise.Model:
    properties:
      brand_id:
        type: string
      id:
        type: string
      name:
        type: string
      series_id:
        type: string
    type: object
  carwise.ModelCreateRequest:
    description: Request body for creating a new model
    properties:
      name:
        example: 320i
        type: string
    required:
    - name
    type: object
  carwise.ModelUpdateRequest:
    description: Request body for updating an existing model
    properties:
      name:
        example: 320i
        type: string
    required:
    - name
    type: object
...
  carwise.SeriesCreateRequest:
    description: Request body for creating a new series
    properties:
      name:
        example: 3 Series
        type: string
    required:
    - name
    type: object
  carwise.SeriesDetail:
    properties:
      id:
        type: string
      models:
        items:
          $ref: '#/definitions/carwise.Model'
        type: array
      name:
        type: string
    type: object
  carwise.SeriesUpdateRequest:
    description: Request body for updating an existing series
    properties:
      name:
        example: 3 Series
        type: string
    required:
    - name
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@carwise.com
    name: API Support
  description: Carwise API documentation with Swagger
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Carwise API
  version: "1.0"
paths:
...
  /brands:
    get:
      consumes:
      - application/json
      description: Retrieve all brands with their associated details
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/carwise.BrandWithDetails'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all brands with details
      tags:
      - Brands - Series - Models
    post:
      consumes:
      - application/json
      description: Create a new brand with the provided details
      parameters:
      - description: Brand creation request
        in: body
        name: brand
        required: true
        schema:
          $ref: '#/definitions/carwise.BrandCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Brand created successfully
          schema:
            additionalProperties:
              allOf:
              - type: string
              - properties:
                  message:
                    type: string
                type: object
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new brand
      tags:
      - Brands - Series - Models
  /brands/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a brand by its ID
      parameters:
      - description: Brand ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Brand deleted successfully
          schema:
            additionalProperties:
              allOf:
              - type: string
              - properties:
                  message:
                    type: string
                type: object
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a brand
      tags:
      - Brands - Series - Models
    put:
      consumes:
      - application/json
      description: Update a brand with the provided details
      parameters:
      - description: Brand ID
        in: path
        name: id
        required: true
        type: string
      - description: Brand update request
        in: body
        name: brand
        required: true
        schema:
          $ref: '#/definitions/carwise.BrandUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Brand updated successfully
          schema:
            additionalProperties:
              allOf:
              - type: string
              - properties:
                  message:
                    type: string
                type: object
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update an existing brand
      tags:
      - Brands - Series - Models
  /brands/{id}/series:
    post:
      consumes:
      - application/json
      description: Create a new series for a specific brand
      parameters:
      - description: Brand ID
        in: path
        name: id
        required: true
        type: string
      - description: Series creation request
        in: body
        name: series
        required: true
        schema:
          $ref: '#/definitions/carwise.SeriesCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Series created successfully
          schema:
            additionalProperties:
              allOf:
              - type: string
              - properties:
                  message:
                    type: string
                type: object
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new series
      tags:
      - Brands - Series - Models
  /brands/{id}/series/{sid}:
    delete:
      consumes:
      - application/json
      description: Delete a series by its ID
      parameters:
      - description: Brand ID
        in: path
        name: id
        required: true
        type: string
      - description: Series ID
        in: path
        name: sid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Series deleted successfully
          schema:
            additionalProperties:
              allOf:
              - type: string
              - properties:
                  message:
                    type: string
                type: object
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a series
      tags:
      - Brands - Series - Models
    put:
      consumes:
      - application/json
      description: Update a series with the provided details
      parameters:
      - description: Brand ID
        in: path
        name: id
        required: true
        type: string
      - description: Series ID
        in: path
        name: sid
        required: true
        type: string
      - description: Series update request
        in: body
        name: series
        required: true
        schema:
          $ref: '#/definitions/carwise.SeriesUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Series updated successfully
          schema:
            additionalProperties:
              allOf:
              - type: string
              - properties:
                  message:
                    type: string
                type: object
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update an existing series
      tags:
      - Brands - Series - Models
  /brands/{id}/series/{sid}/models:
    post:
      consumes:
      - application/json
      description: Create a new model for a specific series
      parameters:
      - description: Brand ID
        in: path
        name: id
        required: true
        type: string
      - description: Series ID
        in: path
        name: sid
        required: true
        type: string
      - description: Model creation request
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/carwise.ModelCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Model created successfully
          schema:
            additionalProperties:
              allOf:
              - type: string
              - properties:
                  message:
                    type: string
                type: object
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new model
      tags:
      - Brands - Series - Models
  /brands/{id}/series/{sid}/models/{mid}:
    delete:
      consumes:
      - application/json
      description: Delete a model by its ID
      parameters:
      - description: Brand ID
        in: path
        name: id
        required: true
        type: string
      - description: Series ID
        in: path
        name: sid
        required: true
        type: string
      - description: Model ID
        in: path
        name: mid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Model deleted successfully
          schema:
            additionalProperties:
              allOf:
              - type: string
              - properties:
                  message:
                    type: string
                type: object
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a model
      tags:
      - Brands - Series - Models
    put:
      consumes:
      - application/json
      description: Update a model with the provided details
      parameters:
      - description: Brand ID
        in: path
        name: id
        required: true
        type: string
      - description: Series ID
        in: path
        name: sid
        required: true
        type: string
      - description: Model ID
        in: path
        name: mid
        required: true
        type: string
      - description: Model update request
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/carwise.ModelUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Model updated successfully
          schema:
            additionalProperties:
              allOf:
              - type: string
              - properties:
                  message:
                    type: string
                type: object
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update an existing model
      tags:
      - Brands - Series - Models
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"